fix docker, migrations

+++ scripts/clean-web.sh
@@ -1,9 +1,11 @@
 #!/usr/bin/env sh
 
-SCRIPT_DIR="$(dirname "$(realpath "$0")")";
-ROOT_DIR="$(dirname "$SCRIPT_DIR")";
+SCRIPTS="$(dirname "$(realpath "$0")")";
+ROOT="$(dirname "$SCRIPTS")";
 
-export ORCHARD_APP_DATA="$ROOT_DIR/App_Data";
+export ORCHARD_APP_DATA="$ROOT/App_Data";
 
-printf "\n[OZDS] Cleaning '%s'...\n" "$ORCHARD_APP_DATA";
-rm -rf "$ORCHARD_APP_DATA";
+if [ -d "$ORCHARD_APP_DATA" ]; then
+  printf "[OZDS] Cleaning '%s'...\n" "$ORCHARD_APP_DATA";
+  rm -rf "$ORCHARD_APP_DATA";
+fi;
